Reorganized data directory.
